+++ contracts/hackatom/src/contract.rs
@@ -2,10 +2,10 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, ResultExt};
 
-use cosmwasm::errors::{unauthorized, NotFound, ParseErr, Result, SerializeErr};
-use cosmwasm::serde::{from_slice, to_vec};
-use cosmwasm::traits::{Api, Extern, Storage};
-use cosmwasm::types::{log, CanonicalAddr, CosmosMsg, Env, HumanAddr, Response};
+use cosmwasm::{
+    from_slice, log, to_vec, unauthorized, Api, CanonicalAddr, CosmosMsg, Env, Extern, HumanAddr,
+    NotFound, ParseErr, Response, Result, SerializeErr, Storage,
+};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InitMsg {
@@ -149,11 +149,8 @@ fn query_verifier<S: Storage, A: Api>(deps: &Extern<S, A>) -> Result<Vec<u8>> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use cosmwasm::mock::{dependencies, mock_env};
-    use cosmwasm::storage::transactional_deps;
-    // import trait to get access to read
-    use cosmwasm::traits::ReadonlyStorage;
-    use cosmwasm::types::coin;
+    // import trait ReadonlyStorage to get access to read
+    use cosmwasm::{coin, dependencies, mock_env, transactional_deps, ReadonlyStorage};
 
     #[test]
     fn proper_initialization() {