+++ contracts/hackatom/src/lib.rs
@@ -2,7 +2,7 @@ pub mod contract;
 
 /** Below we expose wasm exports **/
 #[cfg(target_arch = "wasm32")]
-pub use cosmwasm::exports::{allocate, deallocate};
+pub use cosmwasm::{allocate, deallocate};
 
 #[cfg(target_arch = "wasm32")]
 pub use wasm::{handle, init};
@@ -10,13 +10,13 @@ pub use wasm::{handle, init};
 #[cfg(target_arch = "wasm32")]
 mod wasm {
     use super::contract;
-    use cosmwasm::{exports, imports};
+    use cosmwasm::{do_handle, do_init, do_query, ExternalApi, ExternalStorage};
     use std::ffi::c_void;
 
     #[no_mangle]
     pub extern "C" fn init(env_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
-        exports::do_init(
-            &contract::init::<imports::ExternalStorage, imports::ExternalApi>,
+        do_init(
+            &contract::init::<ExternalStorage, ExternalApi>,
             env_ptr,
             msg_ptr,
         )
@@ -24,8 +24,8 @@ mod wasm {
 
     #[no_mangle]
     pub extern "C" fn handle(env_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
-        exports::do_handle(
-            &contract::handle::<imports::ExternalStorage, imports::ExternalApi>,
+        do_handle(
+            &contract::handle::<ExternalStorage, ExternalApi>,
             env_ptr,
             msg_ptr,
         )
@@ -33,9 +33,6 @@ mod wasm {
 
     #[no_mangle]
     pub extern "C" fn query(msg_ptr: *mut c_void) -> *mut c_void {
-        exports::do_query(
-            &contract::query::<imports::ExternalStorage, imports::ExternalApi>,
-            msg_ptr,
-        )
+        do_query(&contract::query::<ExternalStorage, ExternalApi>, msg_ptr)
     }
 }