+++ contracts/hackatom/tests/integration.rs
@@ -1,9 +1,8 @@
 use std::str::from_utf8;
 
-use cosmwasm::mock::mock_env;
-use cosmwasm::serde::from_slice;
-use cosmwasm::traits::{Api, ReadonlyStorage};
-use cosmwasm::types::{coin, log, CosmosMsg, HumanAddr, QueryResult};
+use cosmwasm::{
+    coin, from_slice, log, mock_env, Api, CosmosMsg, HumanAddr, QueryResult, ReadonlyStorage,
+};
 
 use cosmwasm_vm::testing::{handle, init, mock_instance, query, test_io};
 
@@ -197,7 +196,7 @@ fn passes_io_tests() {
 mod singlepass_tests {
     use super::*;
 
-    use cosmwasm::serde::to_vec;
+    use cosmwasm::to_vec;
     use cosmwasm_vm::call_handle;
     use cosmwasm_vm::testing::mock_instance_with_gas_limit;