+++ packages/std/src/lib.rs
@@ -1,21 +1,39 @@
 // Exposed on all platforms
 
-pub mod encoding;
-pub mod errors;
-pub mod mock;
-pub mod serde;
-pub mod storage;
-pub mod traits;
-pub mod types;
+mod encoding;
+mod errors;
+mod mock;
+mod serde;
+mod storage;
+mod traits;
+mod types;
+
+pub use crate::encoding::Binary;
+pub use crate::errors::{
+    contract_err, dyn_contract_err, invalid, unauthorized, Error, NotFound, NullPointer, ParseErr,
+    Result, SerializeErr,
+};
+pub use crate::mock::{dependencies, mock_env, MockApi, MockStorage};
+pub use crate::serde::{from_slice, to_vec};
+pub use crate::storage::{transactional, transactional_deps};
+pub use crate::traits::{Api, Extern, ReadonlyStorage, Storage};
+pub use crate::types::{
+    coin, log, CanonicalAddr, ContractResult, CosmosMsg, Env, HumanAddr, QueryResult, Response,
+};
 
 // Exposed in wasm build only
 
 #[cfg(target_arch = "wasm32")]
-pub mod exports;
+mod exports;
 #[cfg(target_arch = "wasm32")]
-pub mod imports;
+mod imports;
 #[cfg(target_arch = "wasm32")]
-pub mod memory; // used by exports and imports only
+mod memory; // used by exports and imports only
+
+#[cfg(target_arch = "wasm32")]
+pub use crate::exports::{allocate, deallocate, do_handle, do_init, do_query};
+#[cfg(target_arch = "wasm32")]
+pub use crate::imports::{ExternalApi, ExternalStorage};
 
 // Not exposed